+++ src/godot-builder/index.ts
@@ -1,4 +1,5 @@
 import BuildConfig from './build-config'
 import Image from './image'
+import dockerExport from './docker'
 
-export {BuildConfig, Image}
+export {BuildConfig, Image, dockerExport}