+++ src/godot-builder/input-validation/engine-version.ts
@@ -1,3 +1,0 @@
-const engineVersions = ['4.0', 'latest']
-
-export default engineVersions