+++ CHANGELOG.md
@@ -9,6 +9,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Base docker image
-- Configuration loader
-- Image tag calculator
+- Windows build support
+- Linux build support